fix warning in allocate.c
git-svn-id: http://downloads.open-mesh.org/svn/batman/trunk/batman@168 45894c77-fb22-0410-b583-ff6e7d5dbf6c

+++ allocate.c
@@ -72,7 +72,7 @@ void *debugMalloc(unsigned int length, int tag)
 
 	if (memory == NULL)
 	{
-		fprintf(stderr, "Cannot allocate %u bytes, tag = %d\n", length + sizeof(struct chunkHeader) + sizeof(struct chunkTrailer), tag);
+		fprintf(stderr, "Cannot allocate %u bytes, tag = %d\n", (unsigned int)(length + sizeof(struct chunkHeader) + sizeof(struct chunkTrailer)), tag);
 		exit(1);
 	}
